Fixed overlay flickering while dragging handle
When dragging the handle, the .twentytwenty-overlay previously faded in and out as soon as the mouse left the handle div. Now, the hover class is applied to the .twentytwenty-container directly and gets overwritten at the startSlide() event. Also, v-on:touchstart="startSlide" and v-on:mousedown="startSlide" were moved from the .twentytwenty-overlay to the .twentytwenty-handle to create a user experience more consistent with the original http://zurb.com/playground/twentytwenty.